requirements.txt
sympy>=1.11
matplotlib>=3.6
Pillow>=9.4
